get ride off Summoning class
